refactor: separate multi konva container
